WIP: koji_tag: remove then add in ensure_external_repos
Prior to this change, we would hit an error in the hub because we would
try to add an external repo at the same priority as one that already
exists.

Switch the order of operations so that we delete all the wrong repos
from the hub and then add all the correct repos that are missing.

This change drops our use of editExternalRepos. It could be a tiny bit
faster to use that RPC instead of the pair of remove+add RPCs, but the
edit RPC is currently incompatible with Koji's new merge_mode feature:
https://pagure.io/koji/issue/1857 . We don't support merge_mode in
koji-ansible yet, but we will in a future commit.

This also moves our add/remove RPC loops to make those easier to test
and refactor to multicalls later.